done event and thread

+++ src/libs.rs
@@ -1,5 +1,5 @@
 pub mod frame;
 pub mod render;
 
-pub const ROWS: i32 = 20;
-pub const COLS: i32 = 40;
+pub const ROWS: usize = 20;
+pub const COLS: usize = 40;
